move logic out of App component
