+++ src/App.tsx
@@ -2,36 +2,12 @@ import React from 'react';
 
 import './App.scss';
 
-import postsFromServer from './api/posts';
-import commentsFromServer from './api/comments';
-import usersFromServer from './api/users';
-import { Post } from './types/Post';
-import { User } from './types/User';
-import { PostComment } from './types/PostComment';
 import { PostList } from './components/PostList';
-
-const preparePosts = (
-  posts: Post[],
-  comments: PostComment[],
-  users: User[],
-) => {
-  return posts.map(post => ({
-    ...post,
-    user: users.find(user => user.id === post.userId) || null,
-    comments: comments.filter(comment => comment.postId === post.id),
-  }));
-};
-
-const preparedPosts = preparePosts(
-  postsFromServer,
-  commentsFromServer,
-  usersFromServer,
-);
 
 export const App: React.FC = () => (
   <section className="App">
     <h1 className="App__title">Static list of posts</h1>
 
-    <PostList posts={preparedPosts} />
+    <PostList />
   </section>
 );
